+++ tests/vec3.rs
@@ -195,8 +195,6 @@ macro_rules! impl_vec3_tests {
             assert!(c.cmple(c).all());
             assert!(c.cmpge(c).all());
             assert!(a == a);
-            assert!(a < b);
-            assert!(b > a);
         }
 
         #[test]
@@ -384,9 +382,6 @@ macro_rules! impl_vec3_tests {
             assert_eq!(b, a);
             assert_ne!(a, c);
             assert_ne!(b, c);
-
-            assert!(a > c);
-            assert!(c < a);
         }
 
         #[test]